users bugs - create, destroy

+++ app/admin/users.rb
@@ -54,7 +54,7 @@ ActiveAdmin.register User, :namespace => false do
     column "" do |resource|
       links = ''.html_safe
       links += link_to I18n.t('active_admin.edit'), edit_resource_path(resource), :class => "member_link edit_link"
-      links += link_to I18n.t('active_admin.delete'), resource_path(resource), :method => :delete, :confirm => I18n.t('active_admin.delete_confirmation'), :class => "member_link delete_link"
+      links += link_to I18n.t('active_admin.delete'), resource_path(resource), :method => :delete, :confirm => I18n.t('active_admin.delete_confirmation'), :class => "member_link delete_link" unless current_user.id == resource.id
       links
     end
   end
